move theme scripts to scripts directory
