+++ packages/core/theme-module/scripts/compileTemplate.js
