+++ packages/core/theme-module/scripts/copyThemeFiles.js
