+++ packages/core/theme-module/scripts/ensureDirectoryExists.js
